requirements.txt
flet>=0.80.5
camoufox>=0.4.11
python-dotenv>=1.0.0
aiohttp>=3.9.0